Moved common stuff to SpotifyMusicService
