+++ MeNext.Spotify.Droid.Auth/Transforms/EnumMethods.xml
@@ -1,0 +1,19 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<enum-method-mappings>
+	<!--
+  This example changes the Java method:
+    android.support.v4.app.Fragment.SavedState.writeToParcel (int flags)
+  to be:
+    android.support.v4.app.Fragment.SavedState.writeToParcel (Android.OS.ParcelableWriteFlags flags)
+  when bound in C#.
+
+  <mapping jni-class="android/support/v4/app/Fragment.SavedState">
+    <method jni-name="writeToParcel" parameter="flags" clr-enum-type="Android.OS.ParcelableWriteFlags" />
+  </mapping>
+
+  Notes:
+  - For Java interfaces, use "jni-interface" attribute instead of "jni-class" attribute.
+  - To change the type of the return value, use "return" as the parameter name.
+  - The parameter names will be p0, p1, ... unless you provide JavaDoc file in the project.
+  -->
+</enum-method-mappings>
