+++ MeNext.Spotify.Droid/Resources/values/Strings.xml
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<resources>
+	<string name="library_name">MeNext.Spotify.Droid</string>
+</resources>
