+++ MeNext.Spotify.Droid/SpotifyMusicServiceDroid.cs
@@ -1,0 +1,10 @@
+﻿using System;
+namespace MeNext.Spotify.Droid
+{
+    public class SpotifyMusicServiceDroid// : SpotifyMusicServiceDroid
+    {
+        public SpotifyMusicServiceDroid()
+        {
+        }
+    }
+}
