+++ MeNext.sln
@@ -17,6 +17,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MeNext.Spotify.iOS.Playback
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MeNext.Spotify", "MeNext.Spotify\MeNext.Spotify.csproj", "{FA839F74-9C98-40DB-894A-8F7A9B6D43DB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MeNext.Spotify.Droid.Auth", "MeNext.Spotify.Droid.Auth\MeNext.Spotify.Droid.Auth.csproj", "{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MeNext.Spotify.Droid", "MeNext.Spotify.Droid\MeNext.Spotify.Droid.csproj", "{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -123,6 +127,30 @@ Global
 		{FA839F74-9C98-40DB-894A-8F7A9B6D43DB}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
 		{FA839F74-9C98-40DB-894A-8F7A9B6D43DB}.Debug|iPhone.ActiveCfg = Debug|Any CPU
 		{FA839F74-9C98-40DB-894A-8F7A9B6D43DB}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Release|iPhone.Build.0 = Release|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{7FD4B2B8-DDD5-4CF1-8A4B-3E2C08FE3378}.Debug|iPhone.Build.0 = Debug|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Debug|iPhoneSimulator.ActiveCfg = Debug|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Debug|iPhoneSimulator.Build.0 = Debug|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Release|iPhone.ActiveCfg = Release|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Release|iPhone.Build.0 = Release|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Release|iPhoneSimulator.ActiveCfg = Release|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Release|iPhoneSimulator.Build.0 = Release|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Debug|iPhone.ActiveCfg = Debug|Any CPU
+		{6E5EF8EF-309E-4DF6-86F6-596F9FE16834}.Debug|iPhone.Build.0 = Debug|Any CPU
 	EndGlobalSection
 	GlobalSection(MonoDevelopProperties) = preSolution
 		Policies = $0
